requirements.txt
accelerate==1.12.0
annotated-doc==0.0.4
anyio==4.12.1
certifi==2026.2.25
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.2
cycler==0.12.1
diffusers==0.36.0
exceptiongroup==1.3.1
filelock==3.25.0
fonttools==4.61.1
fsspec==2026.2.0
ftfy==6.3.1
h11==0.16.0
hf-xet==1.3.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.5.0
idna==3.11
importlib_metadata==8.7.1
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.9
lpips==0.1.4
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
open_clip_torch==3.3.0
packaging==26.0
pandas==2.3.3
pillow==12.1.1
psutil==7.2.2
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
regex==2026.2.28
requests==2.32.5
rich==14.3.3
safetensors==0.7.0
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
shellingham==1.5.4
six==1.17.0
sympy==1.14.0
threadpoolctl==3.6.0
timm==1.0.25
tokenizers==0.22.2
torch==2.10.0
torchaudio==2.10.0
torchvision==0.25.0
tqdm==4.67.3
transformers==5.2.0
typer==0.24.1
typer-slim==0.24.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
wcwidth==0.6.0
zipp==3.23.0